Views Not Tables / Speed Increase
Changed information_schema.tables to information_schema.views to adjust for the restructured Dataport. (TODO) There's only one electricity_egauge_minutes table now, so much of that section is no longer necessary; I didn't adjust for that here.

To increase the speed of the converter, the "get buildings present in this table" and "get first and last timestamps for this house in this table" sections now pull the required information from the metadata table rather than the electricity_egauge_minutes table. This requires much less searching, since the metadata table is much smaller. It should remain accurate since the metadata table is refreshed every time the eGauge tables are refreshed.